# Fallback for upsert when arrow cannot compare source rows with target rows