<!-- Fixes https://github.com/apache/iceberg-python/issues/1711 -->

## Rationale for this change

Upsert operations in PyIceberg rely on Arrow joins between source and target rows. However, Arrow Acero cannot compare certain complex types — like `struct`, `list`, and `map` — unless they’re part of the join key. When such types exist in non-join columns, the upsert fails with an error like:

```ArrowInvalid: Data type struct<...> is not supported in join non-key field venue_geo```


This PR introduces a **fallback mechanism**: if Arrow fails to join due to unsupported types, we fall back to comparing only the key columns. Non-key complex fields are ignored in the join condition, but still retained in the final upserted data.

---

### Before

```python
# Fails if venue_geo is a non-key struct field
txn.upsert(df, join_cols=["match_id"])
```

> ❌ ArrowInvalid: Data type struct<...> is not supported in join non-key field venue_geo

---

### After

```python
# Falls back to key-based join and proceeds
txn.upsert(df, join_cols=["match_id"])
```

> ✅ Successfully inserts or updates the record, skipping complex field comparison during join

---

## ✅ Are these changes tested?

Yes:
- A test was added to reproduce the failure scenario with complex non-key fields.
- The new behavior is verified by asserting that the upsert completes successfully using the fallback logic.

---

> ℹ️ **Note**  
> This change does not affect users who do not include complex types in their schemas. For those who do, it improves resilience while preserving data correctness.

---

## Are there any user-facing changes?

Yes — upserts involving complex non-key columns (like `struct`, `list`, or `map`) no longer fail. They now succeed by skipping unsupported comparisons during the join phase.