@koenvo 2025-04-08
### Poor man's benchmark

This compares the performance of the original vs fallback `upsert` logic.  
The "With skips" case simulates situations where non-matching rows can be skipped during comparison.

| Condition      | Original (s) | Fallback (s) | Diff (ms) | Diff (%) |
|----------------|--------------|--------------|-----------|----------|
| Without skips  | 0.727        | 0.724        | -2.73     | -0.38%   |
| With skips     | 0.681        | 0.732        | +51.24    | +7.53%   |

No significant performance regression observed. Fallback behaves as expected.