@Fokko 2025-04-23
Thanks @kevinjqliu for the write-up. I checked locally and also noticed that the fallback has a different behavior:

```
FAILED tests/table/test_upsert.py::test_merge_scenario_skip_upd_row - AssertionError: rows updated should be 1, but got 2
FAILED tests/table/test_upsert.py::test_merge_scenario_date_as_key - AssertionError: rows updated should be 1, but got 2
FAILED tests/table/test_upsert.py::test_merge_scenario_string_as_key - AssertionError: rows updated should be 1, but got 2
FAILED tests/table/test_upsert.py::test_upsert_with_identifier_fields - assert 2 == 1
```

How about reverting my work in https://github.com/apache/iceberg-python/pull/1685? As I wasn't aware of the limitation of Arrow which doesn't support complex types in joins.