@koenvo 2025-04-23
I was already experimenting with creating a mask. So instead of creating an expression (like you referenced to) it will be a mask. 

We probably need to use the result of the join as indices for both tables. This requires to have index columns for both source and target, and do something like take(source_index_column) and take(target_index_column).