@koenvo 2025-04-23
When we don't check if rows are updated, we can probably replace `upsert` by:

```python
with self.transaction() as tx:
    overwrite_mask_predicate = upsert_util.create_match_filter(df, join_cols)
    tx.overwrite(df, overwrite_filter=overwrite_mask_predicate)
```

Only need a way to figure out the updated/inserted rows.