@kevinjqliu 2025-04-24
sadly it looks like comparison functions (`pc.equal`) doesn't work with complex types either, bummer! 

I was able to verify and reproduce it in a jupyter notebook https://gist.github.com/kevinjqliu/33670f6e852fe175724bffd2e2f4cf07