@kevinjqliu 2025-04-25
Leaving this snippet from my stash here for when pyarrow supports comparison of complex types. 
This is a pyarrow compute only implementation

```
    # filters the source_table for matching rows based on join key columns
    match_expr = functools.reduce(operator.and_, [pc.field(col).isin(target_table.column(col)) for col in join_cols])
    matching_source_rows = source_table.filter(match_expr)

    # filters the source_table for non-matching rows based on non join key columns, these need to be updated
    non_match_expr = functools.reduce(operator.or_, [~pc.field(col).isin(target_table.column(col)) for col in non_key_cols])
    rows_to_update_table = matching_source_rows.filter(non_match_expr)
    return rows_to_update_table
```